+++ hamilton/__init__.py
@@ -8,8 +8,9 @@ from __future__ import division, print_function, absolute_import
 from .functions import (conjugate, multiply, norm, normalize, rotate,
                         vector_vector_rotation, from_euler, to_euler,
                         from_matrix, to_matrix, from_axis_angle, to_axis_angle,
-                        from_mirror_plane, reflect, exp, log, log10, logn,
-                        power)
+                        from_mirror_plane, reflect, exp, log, log10, logb,
+                        power, isnan, isinf, isfinite, equal, not_equal,
+                        allclose, isclose, inverse, divide, expb, exp10)
 
 # Get the version
 import os
@@ -20,8 +21,10 @@ with open(os.path.join(os.path.dirname(os.path.abspath(__file__)),
 
 __all__ = ['conjugate',
            'exp',
+           'expb',
+           'exp10',
            'log',
-           'logn',
+           'logb',
            'log10',
            'power',
            'multiply',
@@ -36,5 +39,14 @@ __all__ = ['conjugate',
            'from_matrix',
            'to_matrix',
            'from_axis_angle',
-           'to_axis_angle'
+           'to_axis_angle',
+           'isnan',
+           'isinf',
+           'isfinite',
+           'equal',
+           'not_equal',
+           'allclose',
+           'isclose',
+           'inverse',
+           'divide'
            ]